requirements.txt
Django
psycopg2-binary
Pillow
djangorestframework
drf-yasg==1.20
python-decouple==3.4